% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_te_heatmaps}
\alias{create_te_heatmaps}
\title{Create Transfer Entropy Heatmaps}
\usage{
create_te_heatmaps(
  te_matrices,
  periods = c("Pre_Crisis", "Crisis", "Post_Crisis"),
  quantile = "Q0.1"
)
}
\arguments{
\item{te_matrices}{Named list of transfer entropy matrices}

\item{periods}{Character vector of periods to visualize}

\item{quantile}{Character string specifying quantile to focus on}
}
\value{
Named list of ggplot heatmap objects
}
\description{
Creates heatmap visualizations of transfer entropy matrices to show
the intensity and direction of information flows between countries.
}
\examples{
\dontrun{
heatmaps <- create_te_heatmaps(te_matrices)
gridExtra::grid.arrange(grobs = heatmaps, ncol = 3)
}

}
